refactor: separate view and controller
