+++ src/main/java/baseballgame/BaseballGame.java
@@ -1,7 +1,11 @@
 package baseballgame;
 
+import baseballgame.view.InputView;
+
 import java.util.List;
 import java.util.Objects;
+
+import static baseballgame.view.OutputView.*;
 
 /**
  * @author KSKIM
@@ -9,10 +13,6 @@ import java.util.Objects;
  * @since 2019-11-28
  */
 public class BaseballGame {
-    private static final String MESSAGE_START_GAME = "# 새로운 게임을 시작합니다.";
-    private static final String MESSAGE_EXIT_GAME = "# 게임을 종료합니다.";
-    private static final String MESSAGE_CORRECT_NUMBER = "# 정답입니다.";
-
     private final RandomNumberGenerator randomNumberGenerator;
     private BallCountCalculator ballCountCalculator;
 
@@ -21,7 +21,7 @@ public class BaseballGame {
     }
 
     public void run() {
-        System.out.println(MESSAGE_START_GAME);
+        showStartMessage();
         generateRandomNumber();
         guessNumber();
     }
@@ -36,10 +36,10 @@ public class BaseballGame {
         while(true) {
             Number userNumber = InputView.getNumber();
             BallCount ballCount = ballCountCalculator.getBallCount(userNumber);
-            System.out.println(ballCount);
+            showBallCountMessage(ballCount);
 
             if (ballCount.isStrikeOut()) {
-                System.out.println(MESSAGE_CORRECT_NUMBER);
+                showCorrectMessage();
                 break;
             }
         }
@@ -48,7 +48,7 @@ public class BaseballGame {
     public boolean restart() {
         MenuType selectedMenu = InputView.getMenu();
         if (MenuType.EXIT.equals(selectedMenu)) {
-            System.out.println(MESSAGE_EXIT_GAME);
+            showExitMessage();
             return false;
         }
         return true;
